+++ ricaun.Revit.DB.Shape.Tests/Utils/ShapeExtension.cs
@@ -21,8 +21,29 @@ namespace ricaun.Revit.DB.Shape.Tests
         {
             if (box.Min.IsAlmostEqualTo(other.Min) == false) return false;
             if (box.Max.IsAlmostEqualTo(other.Max) == false) return false;
+            if (box.Transform.AlmostEqual(other.Transform) == false)
+            {
+                //Console.WriteLine(box.Transform.BasisX);
+                //Console.WriteLine(box.Transform.BasisY);
+                //Console.WriteLine(box.Transform.BasisZ);
+                //Console.WriteLine(box.Transform.Origin);
+                //Console.WriteLine(box.Transform.Scale);
 
-            return box.Transform.AlmostEqual(other.Transform);
+                //Console.WriteLine(other.Transform.BasisX);
+                //Console.WriteLine(other.Transform.BasisY);
+                //Console.WriteLine(other.Transform.BasisZ);
+                //Console.WriteLine(other.Transform.Origin);
+                //Console.WriteLine(other.Transform.Scale);
+
+                //Console.WriteLine(box.Transform.BasisX.IsAlmostEqualTo(other.Transform.BasisX));
+                //Console.WriteLine(box.Transform.BasisY.IsAlmostEqualTo(other.Transform.BasisY));
+                //Console.WriteLine(box.Transform.BasisZ.IsAlmostEqualTo(other.Transform.BasisZ));
+                //Console.WriteLine(box.Transform.Origin.IsAlmostEqualTo(other.Transform.Origin));
+                //Console.WriteLine(box.Transform.Scale.AlmostEqual(other.Transform.Scale));
+
+                return false;
+            }
+            return true;
         }
 
         public static bool AlmostEqual(this double value, double other)